Handle SIGSEGV in ParseTest
This commit uses google test's death tests to handle the case where rpng
segfaults while loading a png. Since a segfault crashes the whole
program and not just the failing test, it is inconvenient to obtain the
results of a full execution.

ParseTest is modified to report a failure if the program segfaults (or
quits unexpectedly) while loading a PNG. Because google test's order of
execution of tests is not guaranteed, the ParseTest suite should be run
first with <build-dir>/test --gtest_filter="ParseTest.*". After ensuring
that there are no inputs cause unexpected program termination, the full
test suite may be executed by removing the filter parameter.

Signed-off-by: Rohith Nallandigal <rnallandigal@gmail.com>